# MyApi.Client/OpenAPIs/openapi.yaml
x-generator: NSwag v14.2.0.0 (NJsonSchema v11.1.0.0 (Newtonsoft.Json v13.0.0.0))
openapi: 3.0.0
info:
  title: MyApi
  version: v1
servers:
  - url: http://localhost:5286
paths:
  /weatherforecast:
    get:
      tags:
        - WeatherForecast
      operationId: WeatherForecast_GetWeatherForecast
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherForecast"
  /image/{id}:
    post:
      tags:
        - Image
      operationId: Image_UploadImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          x-position: 1
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                form:
                  type: string
                  format: binary
                  nullable: true
      responses:
        200:
          description: ""
  /persons:
    get:
      tags:
        - Persons
      description: Gets persons
      operationId: Persons_GetPersons
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
          x-position: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 10
          x-position: 2
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsResultOfPerson"
    post:
      tags:
        - Persons
      description: Create a person
      operationId: Persons_CreatePerson
      requestBody:
        x-name: request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonRequest"
        required: true
        x-position: 1
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /persons/{id}:
    get:
      tags:
        - Persons
      description: Get a person by Id
      operationId: Persons_GetPersonById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          x-position: 1
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        404:
          description: ""
    delete:
      tags:
        - Persons
      description: Delete person
      operationId: Persons_DeletePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          x-position: 1
      responses:
        200:
          description: ""
        404:
          description: ""
components:
  schemas:
    WeatherForecast:
      type: object
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
        temperatureC:
          type: integer
          format: int32
        summary:
          type: string
          nullable: true
        temperatureF:
          type: integer
          format: int32
    ItemsResultOfPerson:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        total:
          type: integer
          format: int64
    Person:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        givenName:
          type: string
        middleName:
          type: string
          nullable: true
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
    CreatePersonRequest:
      type: object
      additionalProperties: false
      properties:
        firstName:
          type: string
        givenName:
          type: string
        middleName:
          type: string
          nullable: true
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date